refactored component locations
